LightRelay deployment script improvements
LightRelay should be deployed only for mainnet, in all other cases we
deploy TestRelay.

There is no point for the additional check if LightRelay is already
deployed, this contract does not need any other special treatment
compared to the rest of contracts like Bridge or TBTCVault.

Added Tenderly and Etherscan verification calls.
